Move STX icon from ledger to dashboard folder
